+++ commix.py
@@ -35,6 +35,9 @@ else:
   # Handle target environment that doesn't support HTTPS verification
   ssl._create_default_https_context = _create_unverified_https_context
 
+from urlparse import urlparse
+from os.path import splitext
+
 from src.utils import menu
 from src.utils import logs
 from src.utils import update
@@ -329,6 +332,7 @@ def main():
             if menu.options.os and checks.user_defined_os():
               user_defined_os = settings.TARGET_OS
 
+            # Procedure for target OS identification.
             for i in range(0,len(settings.SERVER_OS_BANNERS)):
               if settings.SERVER_OS_BANNERS[i].lower() in server_banner.lower():
                 found_os_server = True
@@ -351,6 +355,7 @@ def main():
                     if not checks.identified_os():
                       settings.TARGET_OS = user_defined_os
 
+            # Procedure for target server identification.
             found_server_banner = False
             if settings.VERBOSITY_LEVEL >= 1:
               info_msg = "Identifying the target server... " 
@@ -362,11 +367,12 @@ def main():
                 if settings.VERBOSITY_LEVEL >= 1:
                   print "[ " + Fore.GREEN + "SUCCEED" + Style.RESET_ALL + " ]"
                 if settings.VERBOSITY_LEVEL >= 1:
-                  success_msg = "The server was identified as " 
+                  success_msg = "The target server was identified as " 
                   success_msg += server_banner + Style.RESET_ALL + "."
                   print settings.print_success_msg(success_msg)
                 settings.SERVER_BANNER = server_banner
                 found_server_banner = True
+
                 # Set up default root paths
                 if settings.SERVER_BANNERS[i].lower() == "apache":
                   if settings.TARGET_OS == "win":
@@ -382,9 +388,39 @@ def main():
             if not found_server_banner:
               if settings.VERBOSITY_LEVEL >= 1:
                 print "[ " + Fore.RED + "FAILED" + Style.RESET_ALL + " ]"
-              warn_msg = "Heuristics have failed to identify server."
+              warn_msg = "Heuristics have failed to identify target server."
               print settings.print_warning_msg(warn_msg)
+
+            # Procedure for target application identification
+            found_application_extension = False
+            if settings.VERBOSITY_LEVEL >= 1:
+              info_msg = "Identifying the target application ... " 
+              sys.stdout.write(settings.print_info_msg(info_msg))
+              sys.stdout.flush()
+            root, application_extension = splitext(urlparse(url).path)
+            settings.TARGET_APPLICATION = application_extension[1:].upper()
             
+            if settings.TARGET_APPLICATION:
+              found_application_extension = True
+              if settings.VERBOSITY_LEVEL >= 1:
+                print "[ " + Fore.GREEN + "SUCCEED" + Style.RESET_ALL + " ]"           
+                success_msg = "The target application was identified as " 
+                success_msg += settings.TARGET_APPLICATION + Style.RESET_ALL + "."
+                print settings.print_success_msg(success_msg)
+
+              # Check for unsupported target applications
+              for i in range(0,len(settings.UNSUPPORTED_TARGET_APPLICATION)):
+                if settings.TARGET_APPLICATION.lower() in settings.UNSUPPORTED_TARGET_APPLICATION[i].lower():
+                  err_msg = settings.TARGET_APPLICATION + " exploitation is not yet supported."  
+                  print settings.print_critical_msg(err_msg)
+                  raise SystemExit()
+
+            if not found_application_extension:
+              if settings.VERBOSITY_LEVEL >= 1:
+                print "[ " + Fore.RED + "FAILED" + Style.RESET_ALL + " ]"
+              warn_msg = "Heuristics have failed to identify target application."
+              print settings.print_warning_msg(warn_msg)
+
             # Load tamper scripts
             if menu.options.tamper:
               checks.tamper_scripts()